requirements.txt
asgiref==3.9.1
Django==5.2.6
django-appconf==1.1.0
django-filter==25.1
django-htmx==1.26.0
django-select2==8.4.1
et_xmlfile==2.0.0
mysqlclient==2.2.7
openpyxl==3.1.5
sqlparse==0.5.3
